Move device settings to a shared class between daemon and GUI
