+++ src/Mooltipass/MPSettingsBLE.h
@@ -2,14 +2,11 @@
 #define MPSETTINGSBLE_H
 
 #include "MPSettings.h"
+#include "DeviceSettingsBLE.h"
 
-class MPSettingsBLE : public MPSettings
+class MPSettingsBLE : public DeviceSettingsBLE, public MPSettings
 {
     Q_OBJECT
-
-    //MP BLE only
-    QT_SETTINGS_PROPERTY(bool, reserved_ble, false, MPParams::RESERVED_BLE)
-    QT_SETTINGS_PROPERTY(bool, prompt_animation, false, MPParams::PROMPT_ANIMATION_PARAM)
 
 public:
     MPSettingsBLE(MPDevice *parent, IMessageProtocol *mesProt);
@@ -21,16 +18,9 @@ private slots:
     void setSettings();
 
 private:
-    void connectSendParams(WSServerCon* wsServerCon) override;
-    void fillParameterMapping();
+    void connectSendParams(WSServerCon* wsServerCon);
 
-    QByteArray m_settings;
-    QMap<MPParams::Param, int> m_bleByteMapping;
-
-    static constexpr int RESERVED_BYTE = 0;
-    static constexpr int RANDOM_PIN_BYTE = 1;
-    static constexpr int USER_INTERACTION_TIMEOUT_BYTE = 2;
-    static constexpr int ANIMATION_DURING_PROMPT_BYTE = 3;
+    QByteArray m_lastDeviceSettings;
 };
 
 #endif // MPSETTINGSBLE_H
