+++ src/Settings/DeviceSettingsBLE.h
@@ -1,0 +1,31 @@
+#ifndef DEVICESETTINGSBLE_H
+#define DEVICESETTINGSBLE_H
+
+#include "DeviceSettings.h"
+
+class DeviceSettingsBLE : public DeviceSettings
+{
+    Q_OBJECT
+
+    //MP BLE only
+    QT_SETTINGS_PROPERTY(bool, reserved_ble, false, MPParams::RESERVED_BLE)
+    QT_SETTINGS_PROPERTY(bool, prompt_animation, false, MPParams::PROMPT_ANIMATION_PARAM)
+
+public:
+    DeviceSettingsBLE(QObject *parent);
+    virtual ~DeviceSettingsBLE(){}
+
+    enum BLESettingsByte
+    {
+        RESERVED_BYTE = 0,
+        RANDOM_PIN_BYTE = 1,
+        USER_INTERACTION_TIMEOUT_BYTE = 2,
+        ANIMATION_DURING_PROMPT_BYTE = 3
+    };
+
+protected:
+    void fillParameterMapping();
+    QMap<MPParams::Param, int> m_bleByteMapping;
+};
+
+#endif // DEVICESETTINGSBLE_H
